requirements.txt
asgiref==3.11.1
Django==6.0.2
django-jazzmin==3.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.11.0
sqlparse==0.5.5
Pillow>=10.0.0
django-cors-headers>=4.3.0
python-magic>=0.4.27

